+++ src/talos/persistence/models/__init__.py
@@ -11,6 +11,7 @@ from talos.persistence.models.cart_item import CartItem
 from talos.persistence.models.config_override import ConfigOverride
 from talos.persistence.models.listino_item import ListinoItem
 from talos.persistence.models.panchina_item import PanchinaItem
+from talos.persistence.models.storico_ordine import StoricoOrdine
 from talos.persistence.models.vgp_result import VgpResult
 
 __all__ = [
@@ -20,5 +21,6 @@ __all__ = [
     "ConfigOverride",
     "ListinoItem",
     "PanchinaItem",
+    "StoricoOrdine",
     "VgpResult",
 ]